Nifty 50 list
The "Nifty 50 Stock list" included in this repository contains the list of stocks that constituted the Nifty 50 index as of June 9th. Please note that the composition of the Nifty 50 index is subject to periodic changes, and this list may require updating to reflect the most current constituents. It is recommended to verify and update the list regularly to ensure accurate and up-to-date information for any trading or analysis purposes related to the Nifty 50 index.